test: Actualizar pruebas unitarias para manejar excepciones y tipos de coordenadas en servicios de pedido y restaurante

+++ backend/order-service/src/main/java/com/foodtech/application/service/OrderApplicationService.java
@@ -123,7 +123,7 @@ public class OrderApplicationService implements OrderUseCase {
         }
     }
 
-    private boolean isValidCoordinate(Integer coordinate) {
+    private boolean isValidCoordinate(Double coordinate) {
         return coordinate >= 0 && coordinate <= 100;
     }
